py/requirements.txt
requests>=2.25.0
PyYAML>=6.0
psutil>=5.8.0
python-dotenv>=1.0.0
